requirements.txt
openai-whisper==20231117
yt-dlp==2023.12.30
torch==2.1.2
torchaudio==2.1.2
numpy==1.24.4
ffmpeg-python==0.2.0
